initial unit tests with catch2 and hippomocks

+++ main/app_mqtt.c
@@ -292,11 +292,11 @@ bool handle_ota_mqtt_event(esp_mqtt_event_handle_t event)
   return false;
 }
 
-bool getTemperatureValue(int* value, const cJSON* root, const char* tag)
+bool getTemperatureValue(unsigned int* value, const cJSON* root, const char* tag)
 {
   cJSON * object = cJSON_GetObjectItem(root,tag);
   if (object) {
-    *value = object->valuedouble * 10;
+    *value = (unsigned int) object->valuedouble * 10;
     ESP_LOGI(TAG, "%s: %d.%01d", tag, (*value)/10, (*value)%10);
     return true;
   }
